only test async code on python 3.5+
