+++ puresnmp/test/conftest.py
@@ -1,0 +1,6 @@
+import sys
+
+collect_ignore = []
+if sys.version_info < (3, 5):
+    collect_ignore.append('test_aio_pythonized.py')
+    collect_ignore.append('test_aio_raw.py')
